Update simple-animation.md
This is still a draft but I will try to find an online tool to allow embeddable, editable, interactive code in the readme so people can play with the concept before investing time into learning it. I think it would be a way for me to make my learning more active as well.